Shop Fully Implemented

+++ Assets/Scripts/Inventory/Item/InventoryItem.cs
@@ -25,7 +25,7 @@ namespace Project0.Inventories
         [SerializeField] Sprite icon = null;
         [Tooltip("If true, multiple items of this type can be stacked in the same inventory slot.")]
         [SerializeField] bool stackable = false;
-        [SerializeField] float price;
+        [SerializeField] int price;
 
         // STATE
         static Dictionary<string, InventoryItem> itemLookupCache;
